Praline list modal

+++ pages/api/pralinen/[id].js
@@ -5,6 +5,12 @@ export default async function handler(request, response) {
   await dbConnect();
   const { id } = request.query;
 
+  if (request.method === "PUT") {
+    const pralineData = request.body;
+    await Praline.findByIdAndUpdate(id, pralineData);
+    return response.status(200).json({ status: "Praline updated" });
+  }
+
   if (request.method === "DELETE") {
     await Praline.findByIdAndDelete(id);
     return response.status(200).json({ status: "Praline deleted" });
